TSDK-733 Modularize code
